slack-bots/catchup-bot/requirements.txt
slack_bolt
google-genai
apscheduler
python-dotenv